Things can now be added/removed from site

+++ src/things/sunspec/SunSpecThing.cpp
@@ -355,7 +355,7 @@ void SunSpecThing::parseModel(uint16_t modelId, const std::vector<uint16_t>& buf
         for (const auto& v : model.values()) {
             switch (v.first) {
             case DataPoint::totalActiveAcPower:
-                _propertiesSubject.get_subscriber().on_next({{ReadableThingProperty::power, std::get<int32_t>(v.second)}});
+                _propertiesSubject.get_subscriber().on_next({{ReadableThingProperty::power, (double)std::get<int32_t>(v.second)}});
                 break;
             default:
                 break;
